Refactor: Move Parser to tasker package
