+++ src/main/java/tasker/Parser.java
@@ -1,8 +1,16 @@
-package tasker.command;
+package tasker;
 
 import java.time.LocalDateTime;
 import java.time.format.DateTimeParseException;
 
+import tasker.command.AddCommand;
+import tasker.command.ByeCommand;
+import tasker.command.Command;
+import tasker.command.CommandType;
+import tasker.command.DeleteCommand;
+import tasker.command.ListCommand;
+import tasker.command.MarkCommand;
+import tasker.command.UnmarkCommand;
 import tasker.exception.TaskerException;
 import tasker.task.DateTimeTask;
 import tasker.task.Deadline;
