+++ src/main/java/tasker/command/MarkCommand.java
@@ -7,8 +7,8 @@ import tasker.task.TaskList;
 /**
  * Command for marking a task as done.
  */
-class MarkCommand extends IndexCommand {
-    MarkCommand(int index) {
+public class MarkCommand extends IndexCommand {
+    public MarkCommand(int index) {
         super(index);
     }
