+++ src/main/java/tasker/command/UnmarkCommand.java
@@ -7,8 +7,8 @@ import tasker.task.TaskList;
 /**
  * Command for unmarking a task as done.
  */
-class UnmarkCommand extends IndexCommand {
-    UnmarkCommand(int index) {
+public class UnmarkCommand extends IndexCommand {
+    public UnmarkCommand(int index) {
         super(index);
     }
